Add support for the hybi-10 WebSocket protocol.
This protocol version is defined in
http://tools.ietf.org/html/draft-ietf-hybi-thewebsocketprotocol-10
and is incompatible with the currently popular hixie-76 version.

Add code to handle the new handshake and data framing. Still missing
is support for fragmented frames and unit tests. There is probably a
couple of bugs hiding in there, but it was successfully tested with
Chromium 15.0.839.0, which implements hybi-10.