Maintain a connected flag to prevent attribute errors.
When the underlying Request object fires the notifyFinish Deferred's
errback, the _request attribute gets deleted, supposedly to prevent
memory leaks. That's definitely not the best way to handle references
and probably is not necessary at all, but changing that would require
more in-depth investigation.

For now, to prevent errors when the connection is lost and then the
remaining buffers get written to the now nonexistent request, causing
AttributeErrors, keep a flag that gets set when the connection is lost
and makes further writes and connection drop requests get ignored.